Hmmmn... Treating values a little differently in traits, may want to revisit.

+++ app/models/import/repository.rb
@@ -410,11 +410,12 @@ module Import
         end
         meta_rows << row
       end
-      log('slurping traits and associations and all metadata')
+      return if trait_rows.size <= 1
+      log("slurping traits and associations (#{trait_rows.size - 1}) and all metadata (#{meta_rows.size - 1}, total #{trait_rows.size + meta_rows.size - 2})")
       traits_file = Rails.public_path.join("traits_#{@resource.id}.csv")
       meta_traits_file = Rails.public_path.join("meta_traits_#{@resource.id}.csv")
       CSV.open(traits_file, 'w') { |csv| trait_rows.each { |row| csv << row } }
-      CSV.open(Rails.public_path.join("meta_traits.csv"), 'w') { |csv| meta_rows.each { |row| csv << row } }
+      CSV.open(meta_traits_file, 'w') { |csv| meta_rows.each { |row| csv << row } }
       count = TraitBank.slurp_traits(@resource.id)
       log("Created #{count} associations (including metadata).")
       File.unlink(traits_file) if File.exist?(traits_file)
